Update to final result for main repo

+++ src/pages/material-types/list/components/MaterialTypeList.jsx
@@ -1,4 +1,4 @@
-import { Edit, Forbid, More, Unlock } from "@icon-park/react";
+import { Edit, Forbid, More, PreviewOpen, Unlock } from "@icon-park/react";
 import { Button, Dropdown, Space, message } from "antd";
 import React, { useEffect, useRef, useState } from "react";
 import { BaseTable } from "../../../../components/BaseTable";
@@ -6,17 +6,25 @@ import { MaterialTypeModal } from "../../components/MaterialTypeModal";
 import MaterialCategoryApi from "../../../../apis/material-category";
 import { ConfirmDeleteModal } from "../../../../components/ConfirmDeleteModal";
 import confirm from "antd/es/modal/confirm";
-const MaterialTypeList = () => {
+import { PageSize } from "../../../../constants/enum";
+
+const MaterialTypeList = ({ canModify }) => {
   const [loading, setLoading] = useState(false);
   const [showUpdateMaterialTypeModal, setShowUpdateMaterialTypeModal] = useState(false);
   const [materialTypeList, setMaterialTypeList] = useState([]);
+  const [currentPage, setCurrentPage] = useState(1);
+  const [total, setTotal] = useState([]);
 
   const materialTypeRef = useRef();
 
-  const getData = async (keyword) => {
+  const getData = async (keyword, pageIndex) => {
     setLoading(true);
-    const response = await MaterialCategoryApi.getAllMaterialCategory(keyword);
-
+    const response = await MaterialCategoryApi.getAllMaterialCategory(
+      keyword,
+      pageIndex,
+      PageSize.ORDER_LIST
+    );
+    setTotal(response.total || []);
     setMaterialTypeList(response.data);
     setLoading(false);
   };
@@ -25,30 +33,48 @@ const MaterialTypeList = () => {
     getData();
   }, []);
 
+  useEffect(() => {
+    getData(null, 1);
+  }, []);
+  const onPageChange = (current) => {
+    setCurrentPage(current);
+    getData(null, current);
+  };
+
   const getActionItems = (record) => {
     const { isDeleted } = record;
     return [
-      {
-        key: "UPDATE_ROLE",
-        label: "Cập nhật thông tin",
+      // {
+      //   key: "VIEW_DETAIL",
+      //   label: "Xem thông tin chi tiết",
+      //   icon: <PreviewOpen/>,
+      //   onClick: () => {
+      //     materialTypeRef.current = record;
+      //     // setShowUpdateMaterialTypeModal(true);
+      //   },
+      // },
+      canModify.canUpdate && {
+        key: "UPDATE_MATERIAL_TYPE",
+        label: "Xem thông tin chi tiết",
         icon: <Edit />,
         onClick: () => {
           materialTypeRef.current = record;
           setShowUpdateMaterialTypeModal(true);
         },
       },
-      {
+      canModify.canUpdate && {
         key: "SET_STATUS",
         danger: !isDeleted,
         label: !isDeleted ? "Xoá" : "Phục hồi",
         icon: !isDeleted ? <Forbid /> : <Unlock />,
         onClick: () => {
           confirm({
-            title: "Xoá Loại vật liệu",
+            title: "Xoá loại vật liệu",
             content: `Chắc chắn xoá "${record.name}"?`,
             type: "confirm",
-            cancelText: "Cancel",
-            onOk: () => deleteItem(record.id),
+
+            cancelText: "Hủy",
+            onOk: () => deleteMaterialCategory(record.id),
             onCancel: () => {},
             closable: true,
           });
@@ -59,12 +85,14 @@ const MaterialTypeList = () => {
 
   const columns = [
     {
-      title: "ID",
-      dataIndex: "id",
-      key: "id",
-      width: "30%",
-      align: "center",
-      sorter: (a, b) => a.id.localeCompare(b.id),
+      title: "#",
+      dataIndex: "index",
+      key: "index",
+      width: "5%",
+      // align: "center",
+      render: (_, record, index) => {
+        return <span>{index + 1 + (currentPage - 1) * PageSize.ORDER_LIST}</span>;
+      },
     },
     {
       title: "Tên loại vật liệu",
@@ -84,34 +112,51 @@ const MaterialTypeList = () => {
       sorter: (a, b) => a.name.localeCompare(b.name),
     },
     {
-      title: "Tình trạng",
-      dataIndex: "isDeleted",
-      key: "isDeleted",
-      width: "20%",
-      align: "center",
-      render: (_, { isDeleted }) => {
+      title: "Số lượng vật liệu",
+      dataIndex: "quantityMaterial",
+      key: "quantityMaterial",
+      render: (_, record) => {
         return (
-          <span style={{ color: isDeleted ? "#FF0000" : "#29CB00" }}>
-            {isDeleted ? "Không hoạt động" : "Đang hoạt động"}
+          <span
+            onClick={() => {
+              /*showModal(record) */
+            }}
+          >
+            {record.quantityMaterial}
           </span>
         );
       },
-      sorter: (a, b) => a.isDeleted - b.isDeleted,
-      // filter: {
-      //   placeholder: "Chọn trạng thái",
-      //   label: "Trạng thái",
-      //   filterOptions: [
-      //     {
-      //       label: "Đang hoạt động",
-      //       value: false,
-      //     },
-      //     {
-      //       label: "Khóa",
-      //       value: true,
-      //     },
-      //   ],
-      // },
+      sorter: (a, b) => a.quantityMaterial.localeCompare(b.quantityMaterial),
     },
+    // {
+    //   title: "Tình trạng",
+    //   dataIndex: "isDeleted",
+    //   key: "isDeleted",
+    //   width: "20%",
+    //   align: "center",
+    //   render: (_, { isDeleted }) => {
+    //     return (
+    //       <span style={{ color: isDeleted ? "#FF0000" : "#29CB00" }}>
+    //         {isDeleted ? "Không hoạt động" : "Đang hoạt động"}
+    //       </span>
+    //     );
+    //   },
+    //   sorter: (a, b) => a.isDeleted - b.isDeleted,
+    //   // filter: {
+    //   //   placeholder: "Chọn trạng thái",
+    //   //   label: "Trạng thái",
+    //   //   filterOptions: [
+    //   //     {
+    //   //       label: "Đang hoạt động",
+    //   //       value: false,
+    //   //     },
+    //   //     {
+    //   //       label: "Khóa",
+    //   //       value: true,
+    //   //     },
+    //   //   ],
+    //   // },
+    // },
     {
       title: "Thao tác",
       dataIndex: "action",
@@ -120,7 +165,10 @@ const MaterialTypeList = () => {
       align: "center",
       render: (_, record) => {
         return (
-          <Dropdown menu={{ items: getActionItems(record) }}>
+          <Dropdown
+            disabled={!canModify.canCreate || !canModify.canUpdate}
+            menu={{ items: getActionItems(record) }}
+          >
             <Button className="mx-auto flex-center" icon={<More />} />
           </Dropdown>
         );
@@ -132,16 +180,18 @@ const MaterialTypeList = () => {
     getData(value);
   };
 
-  const deleteItem = async (value) => {
-    setLoading(true);
-    const success = await MaterialCategoryApi.deleteMaterialCategory(value);
-    if (success) {
-      message.success("Xoá thành công");
-    } else {
-      message.error("Xoá thất bại");
+  const deleteMaterialCategory = async (value) => {
+    if (window.confirm("Bạn chắc chắn muốn xoá?")) {
+      setLoading(true);
+      const success = await MaterialCategoryApi.deleteMaterialCategory(value);
+      if (success) {
+        message.success("Xoá thành công");
+      } else {
+        message.error("Xoá thất bại");
+      }
+      getData();
+      setLoading(false);
     }
-    getData();
-    setLoading(false);
   };
 
   return (
@@ -149,7 +199,8 @@ const MaterialTypeList = () => {
       <Space className="w-full flex justify-between mb-6">
         <div></div>
         <Button
-          type="primay"
+          disabled={!canModify.canCreate || !canModify.canUpdate}
+          type="primary"
           className="btn-primary app-bg-primary font-semibold text-white"
           onClick={() => setShowUpdateMaterialTypeModal(true)}
         >
@@ -161,7 +212,11 @@ const MaterialTypeList = () => {
         dataSource={materialTypeList}
         columns={columns}
         loading={loading}
-        pagination={false}
+        pagination={{
+          onChange: onPageChange,
+          pageSize: PageSize.ORDER_LIST,
+          total: total,
+        }}
         searchOptions={{
           visible: true,
           placeholder: "Tìm kiếm loại vật liệu...",
